+++ crates/swc_ecma_magicast/src/data.rs
